adding doc

+++ app/controllers/Api.java
@@ -24,6 +24,16 @@ import play.mvc.*;
 import util.Hash;
 
 public class Api extends Controller {
+	
+	public static enum STATUSCODE	{
+		SUCCESS(200), FAIL(400), NOTFOUND(404), FORBIDDEN(403);
+		
+		private int code;
+		private STATUSCODE(int code){
+			this.code = code;
+		}
+		public String toString(){return ""+code;}
+	}
 
 	private static ObjectMapper mapper = new ObjectMapper();
 	private static JsonNodeFactory factory = JsonNodeFactory.instance;
@@ -40,42 +50,47 @@ public class Api extends Controller {
 			if(teacher == null)	{
 				StudentModel stu = StudentModel.findByLoginHashedPass(email, password);
 				if(stu == null)
-					node.put("status", "fail");
+					node.put("status", STATUSCODE.FAIL.toString());
 				else	{
-					node.put("status", "success");
+					node.put("status", STATUSCODE.SUCCESS.toString());
 					node.put("token", ApiTokenModel.generateNewToken());
 				}
 			}else	{
-				node.put("status", "success");
+				node.put("status", STATUSCODE.SUCCESS.toString());
 				node.put("token", ApiTokenModel.generateNewToken());
 			}
 		} catch (NoSuchAlgorithmException e1) {
-			node.put("status", "fail");
+			node.put("status", STATUSCODE.FAIL.toString());
 		}
 		return ok(node);
 	}
 	
 	public static Result addStudent() throws JsonParseException, JsonMappingException, IOException {
-		play.Logger.debug(request().body().asText());
 		/*StudentModel student = mapper.readValue(request().body().asJson(), StudentModel.class);
 		student.insert();*/
 		ObjectNode node = new ObjectNode(factory);
-		node.put("status", 1);
+		node.put("status", STATUSCODE.SUCCESS.toString());
 		return ok(node);
 	}
 
 	public static Result delStudent(String id) {
+		DynamicForm signinForm = form().bindFromRequest();	
+		String token = signinForm.field("t").value();
+		play.Logger.debug(token);
 		ObjectNode node = new ObjectNode(factory);
 		
-		StudentModel student = StudentModel.finder.byId(new ObjectId(id));
-		if(student != null)	{
-			student.delete();
-			node.put("status", 1);
-		} else {
-			node.put("status", 2);
-			node.put("msg", "student not found");
+		if(ApiTokenModel.asToken(token))	{
+			StudentModel student = StudentModel.finder.byId(new ObjectId(id));
+			if(student != null)	{
+				student.delete();
+				node.put("status", STATUSCODE.SUCCESS.toString());
+			} else {
+				node.put("status", STATUSCODE.NOTFOUND.toString());
+				node.put("msg", "student not found");
+			}
+		}else	{
+			node.put("status", STATUSCODE.FORBIDDEN.toString());
 		}
-
 		return ok(node);
 	}
 
@@ -84,14 +99,20 @@ public class Api extends Controller {
 			StudentModel[] all = StudentModel.finder.all().toArray(new StudentModel[0]);
 			ArrayNode ret = new ArrayNode(factory);
 			for(StudentModel student : all)	{
-				String json = mapper.writeValueAsString(student);
-				ret.add(json);
+				ret.add(mapper.convertValue(student, ObjectNode.class));
 			}
-			return ok(ret);
+			ObjectNode node = new ObjectNode(factory);
+			node.put("status", STATUSCODE.SUCCESS.toString());
+			node.put("content", ret);
+			return ok(node);
 		}else	{
 			StudentModel student = StudentModel.finder.byId(new ObjectId(id));
-			String json = mapper.writeValueAsString(student);
-			return ok(json);
+			
+			ObjectNode node = new ObjectNode(factory);
+			node.put("status", STATUSCODE.SUCCESS.toString());
+			node.put("content", mapper.convertValue(student, ObjectNode.class));
+			
+			return ok(node);
 		}
 	}
